re design selected lane header
