+++ scopes/api-reference/renderers/function/function.renderer.module.scss
@@ -42,7 +42,6 @@
 
 .node {
   display: flex;
-  margin-right: 5px;
   font-family: Roboto mono;
   font-size: var(--bit-p-xxs);
   word-break: normal;
