+++ scopes/api-reference/renderers/schema-node-member-summary/variable-node-summary.module.scss
@@ -8,7 +8,6 @@
 
 .node {
   display: flex;
-  margin-right: 5px;
   font-family: Roboto mono;
   font-size: var(--bit-p-xxs);
   word-break: normal;
