+++ scopes/api-reference/renderers/type-array/type-array.renderer.module.scss
@@ -1,6 +1,5 @@
 .node {
   display: flex;
-  margin-right: 5px;
   font-family: Roboto mono;
   font-size: var(--bit-p-xxs);
   word-break: normal;
