+++ scopes/api-reference/renderers/type-intersection/type-intersection.renderer.tsx
@@ -36,7 +36,7 @@ function TypeIntersectionComponent(props: APINodeRenderProps) {
                 metadata={{ [type.__schema]: { columnView: true } }}
               />
               {types.length > 1 && index !== types.length - 1 ? (
-                <div key={`${type.name}-${index}-&`} className={classnames(styles.node, styles.padding2)}>
+                <div key={`${type.name}-${index}-&`} className={classnames(styles.node, styles.separator)}>
                   {'&'}
                 </div>
               ) : null}
@@ -48,7 +48,7 @@ function TypeIntersectionComponent(props: APINodeRenderProps) {
           <div key={`${type.name}-${index}`} className={styles.node}>
             {type.toString()}
             {types.length > 1 && index !== types.length - 1 ? (
-              <div key={`${type.name}-${index}-&`} className={classnames(styles.node, styles.padding2)}>
+              <div key={`${type.name}-${index}-&`} className={classnames(styles.node, styles.separator)}>
                 {'&'}
               </div>
             ) : null}
