+++ scopes/api-reference/renderers/type-union/type-union.renderer.module.scss
@@ -1,12 +1,11 @@
 .node {
   display: flex;
-  margin-right: 5px;
   font-family: Roboto mono;
   font-size: var(--bit-p-xxs);
   word-break: normal;
   color: rgb(57, 115, 0);
 }
 
-.padding2 {
-  padding: 0px 2px;
+.separator {
+  padding: 0px 4px;
 }
